rename mscw_ctm submodules
