+++ tests/plugins/mscw_ctm/test_additional_variables.py
@@ -1,4 +1,4 @@
-from pyaerocom.plugins.mscw_ctm._reader_helpers import (
+from pyaerocom.plugins.mscw_ctm.additional_variables import (
     calc_concNhno3,
     calc_concNnh3,
     calc_concNnh4,
